+++ spec/arachni/session_spec.rb
@@ -222,6 +222,23 @@ describe Arachni::Session do
         end
 
         context 'when a login check is available' do
+            it 'takes into account #check_options' do
+                subject.check_options = {
+                    cookies: {
+                        'custom-cookie' => 'value'
+                    }
+                }
+
+                Arachni::Options.session.check_url = @url
+
+                expect(subject.http).to receive(:request).with(
+                    Arachni::Options.session.check_url,
+                    hash_including( subject.check_options )
+                )
+
+                configured.logged_in?
+            end
+
             context 'and a valid session is available' do
                 it 'returns true' do
                     configured.login